requirements.txt
pyyaml
paho-mqtt
requests
mysql-connector